logger/requirements.txt
django
hazelcast-python-client
